Camera double

+++ cprog_template_win_231017/include/Camera.h
@@ -25,7 +25,7 @@ namespace engine
     private:
         std::shared_ptr<Component> componentToFollow;
         static double distanceMoved;
-        float toMove;
+        double toMove;
     };
 }
 #endif
